Removed the MemGame web-pages
The entire MemGame web-pages were added because somone once told a
client of mine that it was impossible to have a Unity 3D WebGL game that
started up in less than 40 seconds. I placed a WebGL and a WabPlayer
copy of my game to show that it is possible to have a Unity 3D
Web-broswer game that starts up in less than 40 seconds. This hasn't
been needed for a few month now.